move openai-specific modules to openai folder and create agent files
