--- ai/mod.go
@@ -1,1 +1,0 @@
-package ai
