+++ ai/openai/mod.go
@@ -1,0 +1,1 @@
+package openai
