+++ ai/openai/openai_test.go
@@ -1,4 +1,4 @@
-package ai
+package openai
 
 import (
 	"net/http"
